pyproject.toml
[tool.poetry]
name = "discord-selfbot-utilities"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.19.2"
"discord.py" = "^1.7.3"
pytz = "^2021.3"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
bandit = "^1.7.1"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.920"
ossaudit = "^0.5.0"
types-pytz = "^2021.3.3"
types-aiofiles = "^0.8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Customizations
[tool.isort]
src_paths = ["src"]
profile = "black"

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = ["discord.*"]
ignore_missing_imports = true
